Specified Effectve Addresses and refined OpMemory. Added LOAD/STORE.
Split effective addresses away from memory operands. memory operands now
encapsulate an effective address. This allows us to clarify which memory an
address is for (which allows us to define combinations for instructions we
want to support or not). It also allows us to define what happens when there
is a "raw" effective address for an instruction... like should it default to a
particular memory, or any meaning we want. Turns out this is useful.

Wrote some examples for JUMP at the bottom. I yet need to fix every
single instruction with 'mem' in an operand to deal with raw effective
address only or DMEM, SMEM, etc memory operands.